Rename readonly-copy.ts to notifications.ts
